Attempted to add "reload detection".  At least at this point, this seems hopeless -- there is too much varability (due to the game not adding all records to the combatLog.txt file) in which characters appear or don't appear in the combatLog.txt file to make the "signature" based approach for detecting retries to work.
On the other hand, I improved the automatic classification of characters signficantly -- it now checks to see if a character is target hostile creatures more often than not (if so, it is friendly) or targeting friendly creatuers more often than not (of so, it is hostile).

This logic needs more work, though -- the biggest issue is that its scope is limited to a single combat, while it should span multiple combats (even multiple files, if applicable).  I'm not sure what the best strategy is to implement this, so I'm going to go ahead and check in and call it a day.

What's implemented works, and its better than what was used before, so that's something. :)