Major refactoring to attach CombatEvents that *target* a particular character to that characters record.
It doesn't seem to be working, at least not in many cases, but more testing is required.  The application at least runs agains, and the friendly / hostile classificaiton is *much* more accurate than it used to be, although there are still some bugs lurking somewhere in this logic as well.