More work on improving the auto-detect character faction logic.  Its probably as good as it is going to get at this point.
Found and fixed a number of parsing errors, mostly related to failing to extract names from attack / damage / healing events.  In some cases, the names are simply missing, and when those cases are detected the string "NULL" is put into the name field.  Not ideal, because a parsing bug can also cause this situation, but its better than an empty string when the name is legitimately not available in the source data.

There's still issues with how characters are grouped -- far to often, the same "general type" of character isn't being grouped together with other examples of its type.  This seems to be releated to characters whose origin record (the first time the character appeared in the log) was as a target of an attack / damage event.  But I did fix several bugs relating to this very issue and the problem persists...  Its a real pain to troubleshoot.