I may tinker with the autodetect logic some more, but its very close to 100% accurate at properly setting factions *when enough information is available*.  The remaining failures occur when: 1) CHARNAME and company aren't involved in a combat *and* 2) Some of the characters that are in combat have names that include "*CR_*" and some do not *and* 3) Some of the characters that are in combat are unique / semi-unique (that is, there is no opportunity for the player & companions to engage them in combat elsewhere).
When all three conditions are met, the creatures without "*CR_*" in the name get classified as friendly.  This is pretty rare (although it *does* happen), and there doesn't seem to be a good solution for this.

Next up is to add stats to the character detail window -- so if you click on the hyperlink to view "All events related to this character", you get stats for just that character.  I don't think this will be all that difficult, but we'll see.