Lots of changes to the CSV export -- it exports more data now (added friendly name, faction, and combat ID).
Added Combat ID, which segregates CombatEvents by Combat.  The primary purpose is to add <HR> HTML elements to the various WebBrowser windows to indicate that the previous combat has ended and a new combat has begun.  A "# of combats" stat will probably be added, eventually.

Found and fixed some bugs related to To-Hit Target / Critical Confirmation Target calculations -- when a Natural 20 or 1 was rolled, the combatLog.txt doesn't present the Attack Bonus (or Critical HIt Confirmation bonus) in the same line that it reports the die roll.  It *does*, however, still include attack bonus information, but not until several lines later in the same record.  I modified the parsing process to calculate and add target values to attack rolls and critical conformation rolls once attack bonus is available.

Started the long process of adding comments.