Trying (and failing) to improve UI responsiveness when in "CombatEventContainer.Force_Update_Characters_List()".  With large datasets, this can block the UI thread for a good 10+ seconds, but because of where it is used, its hard to make it asynchronious.  If only there was a way to say "Spin the main thread until this returns, then continue execution at this point."
Oh, wait, there *is* -- its async / await.  But when I tried to do that, sometimes the wait finished, and sometimes it didn't.  If I added a timeout, it always returned (eventually), but killed performance.  And besides, when the UI thread was in ".Wait()", it *still blocked the UI thread*.

Ugh.

The alternative is to find a way to cache the Character_List_Item.Get_All_CombatEvents(), because that's the source of the problem -- it gest called lots during VFR (especially) and its very CPU expensive.  But that isn't so easy, because their is only one connection between a CharacterListITem and its children, and (in this case) it goes the wrong way.